refactor: move duplicate code to a helper file
