+++ src/presentation/pages/login/login.spec.tsx
@@ -11,6 +11,7 @@ import {
   AuthenticationSpy,
   ValidationStub,
   SaveAccessTokenMock,
+  Helper,
 } from '@/presentation/test';
 import { InvalidCredentialsError } from '@/domain/errors';
 import { faker } from '@faker-js/faker';
@@ -78,21 +79,6 @@ const populatePasswordField = (
   });
 };
 
-const testStatusForField = (
-  sut: RenderResult,
-  fieldName: string,
-  validationError?: string
-): void => {
-  const emailStatus = sut.getByTestId(`${fieldName}-status`);
-  expect(emailStatus.title).toBe(validationError || 'OK');
-  expect(emailStatus.textContent).toBe(validationError ? '🔴' : '🟢');
-};
-
-const testErrorWrapChildCount = (sut: RenderResult, count: number): void => {
-  const errorWrap = sut.getByTestId('error-wrap');
-  expect(errorWrap.childElementCount).toBe(count);
-};
-
 const testElementExists = (sut: RenderResult, fieldName: string): void => {
   const el = sut.getByTestId(fieldName);
   expect(el).toBeTruthy();
@@ -101,57 +87,49 @@ const testElementText = (sut: RenderResult, fieldName: string, text: string): vo
   const el = sut.getByTestId(fieldName);
   expect(el.textContent).toBe(text);
 };
-const testButtonIsDisabled = (
-  sut: RenderResult,
-  fieldName: string,
-  isDisabled: boolean
-): void => {
-  const button = sut.getByTestId(fieldName) as HTMLButtonElement;
-  expect(button.disabled).toBe(isDisabled);
-};
 
 describe('Login Component', () => {
   afterEach(cleanup);
   test('Should start with initial state', () => {
     const validationError = faker.word.words();
     const { sut } = makeSut({ validationError });
-    testErrorWrapChildCount(sut, 0);
-    testButtonIsDisabled(sut, 'submit', true);
-    testStatusForField(sut, 'email', validationError);
-    testStatusForField(sut, 'password', validationError);
+    Helper.testChildCount(sut, 'error-wrap', 0);
+    Helper.testButtonIsDisabled(sut, 'submit', true);
+    Helper.testStatusForField(sut, 'email', validationError);
+    Helper.testStatusForField(sut, 'password', validationError);
   });
 
   test('Should show emailError if validation fails', () => {
     const validationError = faker.word.words();
     const { sut } = makeSut({ validationError });
     populateEmailField(sut);
-    testStatusForField(sut, 'email', validationError);
+    Helper.testStatusForField(sut, 'email', validationError);
   });
 
   test('Should show passwordError if validation fails', () => {
     const validationError = faker.word.words();
     const { sut } = makeSut({ validationError });
     populatePasswordField(sut);
-    testStatusForField(sut, 'password', validationError);
+    Helper.testStatusForField(sut, 'password', validationError);
   });
 
   test('Should show valid email state if validation succeeds', () => {
     const { sut } = makeSut();
     populateEmailField(sut);
-    testStatusForField(sut, 'email');
+    Helper.testStatusForField(sut, 'email');
   });
 
   test('Should show valid password state if validation succeeds', () => {
     const { sut } = makeSut();
     populatePasswordField(sut);
-    testStatusForField(sut, 'password');
+    Helper.testStatusForField(sut, 'password');
   });
 
   test('Should enable submit button is form is valid', () => {
     const { sut } = makeSut();
     populateEmailField(sut);
     populatePasswordField(sut);
-    testButtonIsDisabled(sut, 'submit', false);
+    Helper.testButtonIsDisabled(sut, 'submit', false);
   });
 
   test('Should show spinner on submit', async () => {
@@ -192,7 +170,7 @@ describe('Login Component', () => {
     jest.spyOn(authenticationSpy, 'auth').mockReturnValueOnce(Promise.reject(error));
     await simulateValidSubmit(sut);
     testElementText(sut, 'main-error', error.message);
-    testErrorWrapChildCount(sut, 1);
+    Helper.testChildCount(sut, 'error-wrap', 1);
   });
 
   test('Should call SaveAccesToken on success', async () => {
@@ -209,7 +187,7 @@ describe('Login Component', () => {
     await simulateValidSubmit(sut);
     await waitFor(() => error);
     testElementText(sut, 'main-error', error.message);
-    testErrorWrapChildCount(sut, 1);
+    Helper.testChildCount(sut, 'error-wrap', 1);
   });
 
   test('Should go to signup page', async () => {
