+++ src/presentation/test/form-helper.tsx
@@ -1,0 +1,27 @@
+import { RenderResult } from '@testing-library/react';
+
+export const testChildCount = (
+  sut: RenderResult,
+  fieldName: string,
+  count: number
+): void => {
+  const el = sut.getByTestId(fieldName);
+  expect(el.childElementCount).toBe(count);
+};
+export const testButtonIsDisabled = (
+  sut: RenderResult,
+  fieldName: string,
+  isDisabled: boolean
+): void => {
+  const button = sut.getByTestId(fieldName) as HTMLButtonElement;
+  expect(button.disabled).toBe(isDisabled);
+};
+export const testStatusForField = (
+  sut: RenderResult,
+  fieldName: string,
+  validationError?: string
+): void => {
+  const fieldStatus = sut.getByTestId(`${fieldName}-status`);
+  expect(fieldStatus.title).toBe(validationError || 'OK');
+  expect(fieldStatus.textContent).toBe(validationError ? '🔴' : '🟢');
+};
